@@ -1,10 +1,11 @@
-# =-
-# (C)opyright Net-ng 2008-2024
+# --
+# Copyright (c) 2008-2025 Net-ng.
+# All rights reserved.
 #
-# This is Net-ng proprietary source code
-# Any reproduction modification or use without prior written
-# approval from Net-ng is strictly forbidden.
-# =-
+# This software is licensed under the BSD License, as described in
+# the file LICENSE.txt, which you should have received as part of
+# this distribution.
+# --
 
 [build-system]
 requires = ['setuptools', 'wheel', 'setuptools_scm']
@@ -14,14 +15,14 @@ build-backend = 'setuptools.build_meta'
 
 [project]
 name = 'nagare-services-webassets'
-dynamic = ['version', 'entry-points']
+dynamic = ['version', 'entry-points', 'scripts']
 description = 'Static assets management'
 authors = [
     {name = 'Net-ng'},
     {email = 'contact@net-ng.com'}
 ]
 license = {file = 'LICENSE.txt'}
-requires-python = '>=3.7'
+requires-python = '>=3.9'
 dependencies = [
     'dukpy',
     'nagare-server'
